+++ src/main/clojure/com/github/clojure_lsp/intellij/file_system.clj
@@ -1,30 +1,35 @@
 (ns com.github.clojure-lsp.intellij.file-system
   (:require
    [clojure.java.io :as io]
-   [com.github.ericdallo.clj4intellij.app-manager :as app-manager])
+   [com.rpl.proxy-plus :refer [proxy+]])
   (:import
    [com.intellij.openapi.project Project]
-   [com.intellij.openapi.vfs VirtualFile]
-   [com.intellij.openapi.vfs VirtualFileManager]))
+   [com.intellij.openapi.vfs LocalFileSystem VirtualFile]))
 
 (set! *warn-on-reflection* true)
 
+(defn temp-path [^Project project]
+  (str (com.intellij.openapi.application.PathManager/getPluginsPath) "/clojure-lsp/cache/" (.getName project)))
+
 (defn create-temp-file ^VirtualFile
   [^Project project ^String path ^String text]
-  (let [temp-file-path (str (com.intellij.openapi.application.PathManager/getPluginsPath) "/clojure-lsp/cache/" (.getName project) path)
-        temp-file (io/file temp-file-path)
-        virtual-file-manager (VirtualFileManager/getInstance)]
-    @(app-manager/write-action!
-      {:run-fn
-       (fn []
-         @(app-manager/execute-command!
-           {:name "Temp file"
-            :group-id "Clojure LSP"
-            :project project
-            :command-fn (fn []
-                          (io/make-parents temp-file-path)
-                          (let [dir (.refreshAndFindFileByNioPath virtual-file-manager (-> temp-file .getParentFile .toPath))
-                                file (.createChildData dir nil (.getName temp-file))]
-                            (.setBinaryContent file (.getBytes text))
-                            true))}))})
-    (.findFileByNioPath virtual-file-manager (.toPath temp-file))))
+  (let [temp-file-path (str (temp-path project) path)
+        temp-file (io/file temp-file-path)]
+    (io/make-parents temp-file-path)
+    (spit temp-file text)
+    (proxy+ [] VirtualFile
+      (getName [_] (.getName temp-file))
+      (getFileSystem [_] (LocalFileSystem/getInstance))
+      (getPath [_] temp-file-path)
+      (isWritable [_] false)
+      (isDirectory [_] false)
+      (isValid [_] true)
+      (getParent [_] nil)
+      (getChildren [_] [])
+      (contentsToByteArray [_] (.getBytes text))
+      (getTimeStamp [_] 0)
+      (getLength [_] (count (.getBytes text)))
+      (refresh [_ _ _ _])
+      (getInputStream [_] (io/input-stream temp-file))
+      (getOutputStream [_ _ _ _] (io/output-stream temp-file))
+      (getModificationStamp [_] -1))))